secure_programming_techniques/hw3/requirements.txt
pycryptodomex==3.17
tk==0.1.0